requirements.txt
# Web scraping and data collection
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data processing and analysis
pandas>=2.0.0
yfinance>=0.2.35

# Data analysis and processing
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.1
textblob>=0.17.1

# Technical analysis
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Report generation
python-docx>=1.0.0
jinja2>=3.1.0

# Utility packages
python-dotenv>=1.0.0
tqdm>=4.65.0

# Add to existing requirements
groq>=0.3.0

# Market Data APIs
pycoingecko>=3.1.0
python-binance>=1.0.17
glassnode>=0.0.2